+++ src/lib/colors/color-decoration.ts
@@ -2,7 +2,8 @@ import {
   Range,
   TextEditorDecorationType,
   Position,
-  window
+  window,
+  DecorationRangeBehavior
 } from 'vscode';
 
 import { generateOptimalTextColor, IDecoration } from '../util/color-util';
@@ -42,8 +43,14 @@ class ColorDecoration implements IDecoration {
   set decoration(deco: TextEditorDecorationType) {
     this._decoration = deco;
   }
-  public constructor(color: Color) {
+
+  get rgb() {
+    return this.color.rgb;
+  }
+
+  public constructor(color: Color, line: number) {
     this.color = color;
+    this.generateRange(line);
   }
   /**
    * Dispose the TextEditorDecorationType
@@ -53,8 +60,10 @@ class ColorDecoration implements IDecoration {
    * @memberOf ColorDecoration
    */
   public dispose(): void {
-    this._decoration.dispose();
-    this.disposed = true;
+    try {
+      this._decoration.dispose();
+      this.disposed = true;
+    } catch (error) {}
   }
   /**
    * Hide the TextEditorDecorationType.
@@ -96,7 +105,8 @@ class ColorDecoration implements IDecoration {
       borderStyle: 'solid',
       borderColor: this.color.toRgbString(),
       backgroundColor: this.color.toRgbString(),
-      color: generateOptimalTextColor(this.color)
+      color: generateOptimalTextColor(this.color),
+      rangeBehavior: DecorationRangeBehavior.ClosedClosed
     });
     this._decoration = backgroundDecorationType;
   }