+++ src/lib/colors/strategies/argb-strategy.ts
@@ -1,4 +1,4 @@
-import Color from './../color';
+import Color from '../color';
 import ColorExtractor, { IColorStrategy } from '../color-extractor';
 import { LineExtraction, DocumentLine } from '../../util/color-util';
 import { EOL, HEXA_VALUE } from '../../util/regexp';
@@ -7,25 +7,25 @@ const HEXA_PREFIX = '(?:#|0x)';
 export const REGEXP = new RegExp(`(${HEXA_PREFIX}(?:${HEXA_VALUE}{3,4}|${HEXA_VALUE}{6}|${HEXA_VALUE}{8}))${EOL}`, 'gi');
 export const REGEXP_ONE = new RegExp(`^(${HEXA_PREFIX}(?:${HEXA_VALUE}{3,4}|${HEXA_VALUE}{6}|${HEXA_VALUE}{8}))${EOL}`, 'i');
 
-class HexaColorExtractor implements IColorStrategy {
-  public name: string = 'HEXA';
+class ARGBColorExtractor implements IColorStrategy {
+  public name: string = 'ARGB';
 
   private extractRGBValue(value): number[] {
     let rgb: any = /#(.+)/gi.exec(value);
     if (rgb[1].length === 3 || rgb[1].length === 4) {
-      return rgb[1].split('').slice(0, 3).map(_ => parseInt(_ + _, 16));
+      return rgb[1].split('').slice(1, 4).map(_ => parseInt(_ + _, 16));
     }
-    rgb = rgb[1].split('').slice(0, 6).map(_ => parseInt(_, 16));
+    rgb = rgb[1].split('').slice(2, 8).map(_ => parseInt(_, 16));
     return [16 * rgb[0] + rgb[1], 16 * rgb[2] + rgb[3], 16 * rgb[4] + rgb[5]];
   }
   private extractAlphaValue(value): number {
     let rgb: any = /#(.+)/gi.exec(value);
     if (rgb[1].length === 4) {
-      let alpha = rgb[1][3];
+      let alpha = rgb[1][0];
       return (parseInt(`${alpha}${alpha}`, 16) / 255);
     }
     if (rgb[1].length === 8) {
-      let alpha = rgb[1].slice(6, 8);
+      let alpha = rgb[1].slice(0, 2);
       return (parseInt(alpha, 16) / 255);
     }
     return 1;
@@ -54,5 +54,5 @@ class HexaColorExtractor implements IColorStrategy {
   }
 }
 
-ColorExtractor.registerStrategy(new HexaColorExtractor());
-export default HexaColorExtractor;
+ColorExtractor.registerStrategy(new ARGBColorExtractor());
+export default ARGBColorExtractor;