+++ src/lib/variables/variable-decoration.ts
@@ -2,7 +2,8 @@ import {
   Range,
   TextEditorDecorationType,
   Position,
-  window
+  window,
+  DecorationRangeBehavior
 } from 'vscode';
 import { generateOptimalTextColor, IDecoration } from '../util/color-util';
 import Variable from './variable';
@@ -43,6 +44,10 @@ class VariableDecoration implements IDecoration {
   }
   set decoration(deco: TextEditorDecorationType) {
     this._decoration = deco;
+  }
+
+  get rgb() {
+    return this.variable.color.rgb;
   }
 
   public constructor(variable: Variable, line: number) {
@@ -110,7 +115,8 @@ class VariableDecoration implements IDecoration {
         borderStyle: 'solid',
         borderColor: this.variable.color.toRgbString(),
         backgroundColor: this.variable.color.toRgbString(),
-        color: generateOptimalTextColor(this.variable.color)
+        color: generateOptimalTextColor(this.variable.color),
+        rangeBehavior: DecorationRangeBehavior.ClosedClosed
       });
       this._decoration = backgroundDecorationType;
     }